requirements.txt
argon2-cffi==19.1.0
astroid==2.3.1
autopep8==1.4.4
bcrypt==3.1.7
beautifulsoup4==4.8.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-atom==0.16.3
django-bootstrap4==1.0.1
django-braces==1.13.0
django-cors-headers==3.1.1
django-countries==5.5
django-crispy-forms==1.8.0
django-debug-toolbar==2.0
django-localflavor==2.2
django-rest-auth==0.9.5
djangorestframework==3.10.3
Faker==2.0.2
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
misaka==2.1.1
oauthlib==3.1.0
Pillow==6.2.0
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.2
python-dateutil==2.8.0
python-decouple==3.1
python-stdnum==1.12
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.4
spotipy==2.4.4
sqlparse==0.3.0
stripe==2.37.2
text-unidecode==1.3
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
